+++ src/miranda/convert/_data_corrections.py
@@ -52,7 +52,9 @@ def load_json_data_mappings(project: str) -> dict[str, Any]:
     """
     data_folder = Path(__file__).resolve().parent / "data"
 
-    if project == "ghcnd":
+    if project == "canhomt_dly":
+        metadata_definition = json.load(data_folder.joinpath("eccc-canhomt_cf_attrs.json").open("r"))
+    elif project == "ghcnd":
         metadata_definition = json.load(data_folder.joinpath("ghcnd_cf_attrs.json").open("r"))
     elif project == "ghcnh":
         metadata_definition = json.load(data_folder.joinpath("ghcnh_cf_attrs.json").open("r"))