+++ templates/eccc-canhomt_daily.py
@@ -2,43 +2,43 @@ import logging
 from pathlib import Path
 import shutil
 from dask.diagnostics import ProgressBar
+from pint.facets.numpy.numpy_func import out_unit
 
 from miranda.convert.stationdata import convert_statdata_bychunks, q_flag_dict
-from miranda.ghcn import download_ghcn
+from miranda.eccc._homogenized import download_canhomt
 import xarray as xr
 
 
 def main():
 
     logging.basicConfig(level=logging.INFO)
-    working_folder = Path.home().joinpath("ghcnd")
-    start_year = 1981
-    end_year = 2020
+    working_folder = Path.home().joinpath("CanHomTV4/CanHomT_dlyV4")
 
-    lon_bnds = [-76, -75]
+    lon_bnds = [-80, -75]
     lat_bnds = [44, 45]
 
     n_stations = 100
+
+    start_year = 1800
+    end_year = None
+
     update_raw = True
 
     # download station data
-    download_ghcn(
-        project="ghcnd",
+    download_canhomt(
+        project="canhomt_dly",
         working_folder=working_folder,
-        lon_bnds=lon_bnds,
-        lat_bnds=lat_bnds,
         update_raw=update_raw,
     )
 
     # convert ghcn data by chunks of n_stations
     convert_statdata_bychunks(
-        project="ghcnd",
+        project="canhomt_dly",
         working_folder=working_folder,
         lon_bnds=lon_bnds,
         lat_bnds=lat_bnds,
         start_year=start_year,
         end_year=end_year,
-        update_from_raw=update_raw,
         n_stations=n_stations,
         n_workers=6,
     )
@@ -49,12 +49,12 @@ def main():
         out_folder.mkdir(parents=True, exist_ok=True)
         if out_folder.exists():
             shutil.rmtree(out_folder)
-        in_zarrs = sorted(var1.glob("*.zarr"))
+        inzarrs = sorted(var1.glob("*.zarr"))
         ds = xr.concat(
-            [xr.open_zarr(z, decode_timedelta=False) for z in in_zarrs], dim="station"
+            [xr.open_zarr(z, decode_timedelta=False) for z in inzarrs], dim="station"
         ).sortby(["station", "time"])
         out_zarr = out_folder.joinpath(
-            f'{var1.name}_day_NOAA_ghcnd_{ds.time[0].dt.strftime("%Y%m%d").values}-{ds.time[-1].dt.strftime("%Y%m%d").values}.zarr'
+            f'{var1.name}_day_GovCan_CamHomT_v4_{ds.time[0].dt.strftime("%Y%m%d").values}-{ds.time[-1].dt.strftime("%Y%m%d").values}.zarr'
         )
         # remove encoding
         for c in ds.coords: